Port ExportSnapshot/LoadAnimations onto GumServiceSkiaBase (#4460)
Extract ExportSnapshot/LoadAnimations/LoadAnimationsFromProvider/
ElementNameFromPath from MonoGameGum's GumService into new
GumCommon/HostServices classes (GumSnapshotExporter,
GumAnimationLoader), following the #4452/#4459 extraction pattern, so
GumService and GumServiceSkiaBase compose one implementation instead
of duplicating it. ExtractUnresolvedTextures becomes an injectable
seam (null on non-XNALIKE hosts) instead of a static partial method,
since GumSnapshotExporter is a separate class in a separate assembly.

Also wires GumServiceSkiaBase.EnableHotReload's loadAnimationsFromProvider
to the real loader instead of the prior no-op, fixing the gap the
class's own doc comment already flagged as a follow-up.
