+++ GumCommon/AssemblyInfo.cs
@@ -3,6 +3,7 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("MonoGameGum.Tests")]
 [assembly: InternalsVisibleTo("MonoGameGum.Tests.V3")]
 [assembly: InternalsVisibleTo("RaylibGum.Tests")]
+[assembly: InternalsVisibleTo("SkiaGum.Tests")]
 
 // MonoGameGum, KniGum, and FnaGum each link in CustomSetPropertyOnRenderable.cs,
 // which writes to internal members on GraphicalUiElement (e.g., IsFontDirty.set).
